refactor into separate package
